+++ projects/angular-sdk/src/public-api.ts
@@ -4,7 +4,7 @@
 
 // SERVICES (only specific elements)
 // export * from './services/hopper-events.service';
-export * from './services/hopper-proxy.service';
+// export * from './services/hopper-proxy.service';
 
 // DIRECTIVES (only specific elements)
 // export * from './directives/hopper-events.directive';
@@ -13,11 +13,10 @@ export * from './services/hopper-proxy.service';
 export * from './components/cfar-offer-banner/cfar-offer-banner.component';
 export * from './components/cfar-offer-banner-large/cfar-offer-banner-large.component';
 export * from './components/cfar-offer-dialog/cfar-offer-dialog.component';
-export * from './components/cfar-exercise-dialog/cfar-exercise-dialog.component';
+export * from './components/cfar-exercise-flow/cfar-exercise-flow.component';
 // export * from './components/event-button/event-button.component';
 
 // UTILIY CLASSES (only specific elements)
-export * from './utils/api-translator.utils';
 export * from './utils/dialog.utils';
 
 // MODULES (only specific elements)