+++ src/app/pages/pages.module.ts
@@ -15,7 +15,7 @@ import { NotFoundPageComponent } from './not-found/not-found.component';
 import { CfarOfferDialogPageComponent } from './guides/cfar-offer-dialog/cfar-offer-dialog.component';
 import { PagesRoutingModule } from './pages-routing.module';
 import { CfarOfferBannerPageComponent } from './guides/cfar-offer-banner/cfar-offer-banner.component';
-import { CfarExerciseDialogPageComponent } from './guides/cfar-exercise-dialog/cfar-exercise-dialog.component';
+import { CfarExerciseFlowPageComponent } from './guides/cfar-exercise-flow/cfar-exercise-flow.component';
 import { GuidesPageComponent } from './guides/guides.component';
 import { SharedModule } from '../shared/shared.module';
 import { MatCardModule } from '@angular/material/card';
@@ -30,7 +30,7 @@ import { HighlightJsModule } from 'ngx-highlight-js';
     CfarOfferDialogPageComponent,
     CfarOfferBannerPageComponent,
     CfarOfferBannerLargePageComponent,
-    CfarExerciseDialogPageComponent
+    CfarExerciseFlowPageComponent
   ],
   imports: [
     // Angular Module