+++ src/app/shared/components/layout/layout.component.ts
@@ -25,7 +25,7 @@ export class LayoutComponent implements OnInit, OnDestroy {
     this.components.push({ name: 'CFAR Offer Banner', link: 'cfar-offer-banner' });
     this.components.push({ name: 'CFAR Offer Banner Large', link: 'cfar-offer-banner-large' });
     this.components.push({ name: 'CFAR Offer Dialog', link: 'cfar-offer-dialog' });
-    this.components.push({ name: 'CFAR Exercise Dialog', link: 'cfar-exercise-dialog' });
+    this.components.push({ name: 'CFAR Exercise Flow', link: 'cfar-exercise-flow' });
     
     // Working items
     if (!environment.production) {